+++ docs/versionhistory.rst
@@ -10,7 +10,10 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Added the ``allow_duplicate_keys`` parameter to :class:`CBORDecoder`, :func:`load` and
   :func:`loads` (default: ``True``). When set to ``False``, a :exc:`CBORDecodeError` is raised
   upon encountering a duplicate key within the same map.
-  (`#283 <https://github.com/agronholm/cbor2/pull/283>`_)
+  (`#283 <https://github.com/agronholm/cbor2/issues/283>`_)
+- Added support for decoding from any object supporting the buffer API (e.g. ``memoryview`` or
+  ``bytearray``) in addition to ``bytes``
+  (`#297 <https://github.com/agronholm/cbor2/issues/297>`_)
 
 **6.0.1** (2026-04-29)